Refactor values tests.
